Fix nao calculando data correta na virada do mes. fix mes dia invertido para brasil no counselor

+++ src/utils/LongDate.java
@@ -99,7 +99,7 @@ public class LongDate {
         }
     }
 
-    public long getDaysDiffToNow() {
+    public int getDaysDiffToNow() {
         return getDaysDiff(getDateLongAsDate(), Calendar.getInstance().getTime());
     }
